requirement.txt
attrs==23.2.0
decorator==5.1.1
Django==3.2.4
django-cors-headers==4.3.1
django-user-agents==0.4.0
djangorestframework==3.12.4
oauth2client==4.1.3
oauthlib==3.2.2
PyJWT==1.4.0
pytz==2021.1
requests==2.32.2
requests-oauthlib==1.3.1
